Fix PHP 8 fatal errors and deprecations in the web code
- explorer/tx.php: 'continue' outside a loop is a fatal compile error in
  PHP 8; show 'Transaction not found' instead.
- site/memcached.php: the (real) cast was removed in PHP 8.
- exchange2/CExchange*: classes with abstract methods must be abstract.
- Replace strftime() (deprecated since 8.1, removed in 9) with date(),
  (double)/(integer) casts and doubleval() with (float)/(int)/floatval(),
  and drop curl_close() calls (a no-op since PHP 8.0, deprecated in 8.5).

+++ web/yaamp/core/backend/bench.php
@@ -58,7 +58,7 @@ function BenchUpdateChips()
                 ':algo' => $bench->algo,
                 ':chip' => $chip
             ));
-            $avg         = (double) $rates['avg'];
+            $avg         = (float) $rates['avg'];
             $cnt         = intval($rates['cnt']);
             if ($cnt > 250) {
                 $bench->delete();
